iter_repo_rpms fix, add rpm.__repr__ and rpm.buildtup
These are some changes I apparently never got around to pushing! One's a
small fix to avoid typerrors if you call iter_repo_rpms with a single
path instead of a list of paths, and the other adds a nicer repr() for
rpm objects and a rpm.buildtup() function, which is handy.. sometimes?