Handle signature tags better
This adds sigtagtbl (so we have correct tag info for the signature
header) and some other bits and pieces needed to correctly poke around
in the signature header.

Oh, also: add TagEntry and use OrderedDict to store tag information so
we can get it back out in the same order it's in the file.