check-payload-headers: fix comparison order
It turns out that the ordering of the archive entries doesn't reliably
match the file entries; RPM basically just unpacks the archive and then
applies the header information to the given path(s).

This fixes check-payload-headers so that it iterates through the payload
entries and compares the metadata from each of those to the
corresponding header metadata by matching the full file paths.

As at the time of this commit I've run this comparison for each file in
the payload of 400,000 RPMs from CentOS 7, RHEL 7, and Fedora 25 through
29, and found no problems/mismatches, so that's good!

(This also confirms my suspicion that the cpio headers are totally
redundant; rpm2cpio could generate cpio headers from the rpm header
metadata with no loss of data.)