Add experimental DINO code
So DINO is my code-name (pet name?) for a reworked repository/package
format. It's heavily inspired by RPM, ELF, git packfiles, and squashfs.

The 'dino' python module is experimental - somehow both half-baked and
over-engineered, which is impressive - and the file format and constants
and everything are all basically one step past cocktail-napkin sketches.

The `mkdino.py` script will take a bunch of directories full of RPMs and
cram them into a bunch of packfiles that end with '.dino'. Neat trick,
but there isn't even a way to get data back _out_ of the packfile yet.

I could write a lot more about this here, but I should probably save
that energy for writing documentation instead of extremely long commit
messages.