rpmtoys.rpm: more tag iterators, better names
Added .iterdigests(), .iterflags(), .iterverifyflags(), and
.iterfileinfo() to iterate through individual per-file flags or a set of
them.

Also renamed .payloadinfo() to .iterrpmfiles(), since it doesn't
actually touch the payload at all, it just iterates through the RPM
per-file tags and gives you rpmfile objects.

Oh also, rpm.srctup is an alias for rpm.buildtup. Handy for when you
want to know the name of the .src.rpm a package came from. (for example,
did you know that `blas` is actually a subpackage of `lapack`, or that
`minizip-compat` is a supbackage of `zlib`, unrelated to `minizip`? What
fun!)