Woooo, extracting files from .dino RPM archives works
I should break this up into smaller commits but I wanted to make sure
this got pushed somewhere before I broke it somehow.

the DINORPMArchive object can now do .write_rpm(key, outf) to extract an
RPM (given by its key in the index) to a file object.

the resulting RPM is _almost_ identical to the original one - the
headers are identical and the _uncompressed_ archive is the same, but
because of some unpredictable changes in compression algorithms and such
we cannot always reproduce the exact original _compressed_ payload.
But hey, if the uncompressed payloads are identical and we have a good
signature for one, doesn't that mean we're good enough? I think so.