pkgtup: add .fromstr() and .match()
pkgtup.fromstr() is a function that looks at a package name string and
figures out whether to parse it as a NEVRA or ENVRA. You'll still
probably get some garbage if you pass it a non-package-name but.. who
can tell?

and .match is a function that compares each part of the pkgtup that's
set - not None, not '' - and returns True if they all match, and False
if they don't.

We'll probably want a better parse function that handles things like
'glibc.i686' correctly to make .match() useful but hey... not now. It's
10:30pm and I been drinkin', okay