mkdino: real CLI parsing
Working on a proper CLI here! All of this may change but there's four
basic commands - build, list, info, and extract - which can be used to
build new DINO repos, list the contents of a packfile, get more info
about the packfile, or (in theory - I haven't written that bit yet)
extract RPM headers or package files or payloads from the packfile.

yayyyyyy